windows clients:  fix mouse spinning bug when the game window is moved left or above the default monitor (negative window position).  noticed in a recent darkplaces commit: http://svn.icculus.org/twilight?view=rev&revision=12358
git-svn-id: file:///svn/uhexen2/code/trunk@6029 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/gl_vidnt.c
@@ -1624,8 +1624,8 @@ static LRESULT WINAPI MainWndProc (HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM l
 		// ignore when fullscreen
 		if (modestate == MS_FULLDIB)
 			break;
-		window_x = (int) LOWORD(lParam);
-		window_y = (int) HIWORD(lParam);
+		window_x = (short) LOWORD(lParam);
+		window_y = (short) HIWORD(lParam);
 		VID_UpdateWindowStatus ();
 		break;
